requirements.txt
Flask>=2.3.0
Flask-SocketIO>=5.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
watchdog>=3.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0

